Hardcode GHC version for now
The GHC version depends on the Stack resolver defined in the Makefile
and would thus have to be manually updated (for now).

The issue with computing it dynamically (though stack ghc) is that
stack.yaml has to be generated dynamically and that stack.yaml depends
on the kics2 runtime packages, which in turn depend on
Installation.curry/.hs, which depends on the GHC versions (thereby
causing a cyclic dependency).