Remove meta-generation for now
One issue is that we always need to run generation a second time after
invoking ./configure since the ninja_log hasn't cached the command line
for the generation of build.ninja yet (and therefore considers the build
generator to be dirty).

Another issue is that we have to pass all parameters explicitly,
resulting in more duplication.